Ensure stable URLs even on fd.o
On Freedesktop, we don't control the server configuration, so we switch
to using hash-urls on this particular service. This allows people to
still link to individual pages without annoying and unexpected 404
errors.
For every other configuration, we use regular URLs.
This change also adds a nice helper script to build the app for the
right configuration.